+++ lib/streamlit/elements/lib/policies.py
@@ -111,8 +111,6 @@ def check_cache_replay_rules() -> None:
     function to check for those as well. And rename it to check_widget_usage_rules.
     """
     if runtime.exists():
-        # from streamlit.runtime.scriptrunner.script_run_context import get_script_run_ctx
-
         ctx = get_script_run_ctx()
         if ctx and ctx.disallow_cached_widget_usage:
             from streamlit import exception
@@ -120,3 +118,31 @@ def check_cache_replay_rules() -> None:
             # We use an exception here to show a proper stack trace
             # that indicates to the user where the issue is.
             exception(CachedWidgetWarning())
+
+
+_fragment_writes_widget_to_outside_error = (
+    "Fragments cannot write to elements outside of their container."
+)
+
+
+def check_fragment_path_policy(dg: DeltaGenerator):
+    ctx = get_script_run_ctx()
+    # Check is only relevant for fragments
+    if ctx is None or ctx.current_fragment_id is None:
+        return
+
+    current_fragment_delta_path = ctx.current_fragment_delta_path
+    current_cursor = dg._active_dg._cursor
+    if current_cursor is None:
+        return
+
+    current_cursor_delta_path = current_cursor.delta_path
+
+    # the elements delta path cannot be smaller than the fragment's delta path if it is inside of the fragment
+    if len(current_cursor_delta_path) < len(current_fragment_delta_path):
+        raise StreamlitAPIException(_fragment_writes_widget_to_outside_error)
+
+    # all path indices of the fragment-path must occur in the inner-elements delta path, otherwise it is outside of the fragment container
+    for index, path_index in enumerate(current_fragment_delta_path):
+        if current_cursor_delta_path[index] != path_index:
+            raise StreamlitAPIException(_fragment_writes_widget_to_outside_error)
